+++ website/src/logos/index.ts
@@ -2,6 +2,7 @@ export * from './AlgoliaLogo';
 export * from './AnthropicLogo';
 export * from './BoltLogo';
 export * from './BuilderLogo';
+export * from './CodeRabbitLogo';
 export * from './DailyDevLogo';
 export * from './DigitalOceanLogo';
 export * from './HdmLogo';